arreglados errores con eliminar y modificar

+++ Asignaturas/views.py
@@ -18,35 +18,62 @@ def inicio(request):
 def tablaAsignaturas(request):
 	"""Toma las asignaturas de la base de datos y las carga en la tabla."""
 
-	materias = Asignatura.objects.all()
 	user = request.user
+	prof = Profesor.objects.get(user = user)
+	dept = prof.departamento.codigo
 
+	materias = Asignatura.objects.filter(departamento = dept).all()
+	
 	if request.method == 'POST':
 
 		if ((request.POST.get('modo')) == "Eliminar"):
-			item = Asignatura.objects.filter(codigo = request.POST.get('item_id'))       
-			item.delete()
-			materias = Asignatura.objects.all()
+
+			item = Asignatura.objects.filter(codigo = request.POST.get('item_id')) 
+			if (Asignatura.objects.filter(codigo = request.POST.get('item_id')).first() != None):  
+				itemAsig = Asignatura.objects.filter(codigo = request.POST.get('item_id')).first().departamento.codigo
+				# Chequeamos si el usuario autentica es del mismo dpto que desea eliminar
+				prof = Profesor.objects.get(user = user)
+				dept = prof.departamento.codigo
+				if (dept == itemAsig):
+					item.delete()
+					materias = Asignatura.objects.filter(departamento = dept).all()
+
 		elif ((request.POST.get('modo')) == "Modificar"):
-			return redirect('/modificar-asignatura/{}'.format(request.POST.get('item_id')))
+			item = Asignatura.objects.filter(codigo = request.POST.get('item_id')) 
+			if (Asignatura.objects.filter(codigo = request.POST.get('item_id')).first() != None):  
+				itemAsig = Asignatura.objects.filter(codigo = request.POST.get('item_id')).first().departamento.codigo
+				# Chequeamos si el usuario autentica es del mismo dpto que desea eliminar
+				prof = Profesor.objects.get(user = user)
+				dept = prof.departamento.codigo
+				if (dept == itemAsig):
+
+					return redirect('/modificar-asignatura/{}'.format(request.POST.get('item_id')))
+
 		else:
 			return render(request, 'Asignaturas/tablaAsignaturas.html', {'materias': materias})
 
 
 	elif request.method=='GET':
-		materias = Asignatura.objects.all()
+		materias = Asignatura.objects.filter(departamento = dept).all()
 
 	return render(request, 'Asignaturas/tablaAsignaturas.html', {'materias': materias})
 
 def modificarAsignatura(request, codigo):
 	"""Busca la materia en la base de datos y llena el formulario con los datos"""
 	user = request.user
+	prof = Profesor.objects.get(user = user)
+	dept = prof.departamento.codigo
 	materia = Asignatura.objects.get(codigo=codigo)
 	if request.method == 'POST':
 		form = RegistrarMatForm(user, request.POST, instance=materia)
 		if form.is_valid():
-			form.save()
-			return redirect('/tabla-asignaturas/')
+			codigo = form.cleaned_data['codigo'][0]+form.cleaned_data['codigo'][1]
+
+			if (codigo == dept):
+				form.save()
+				return redirect('/tabla-asignaturas/')
+			else:
+				return render(request, 'Asignaturas/modificarAsignatura.html', {'form':form})
 		else:
 			return render(request, 'Asignaturas/modificarAsignatura.html', {'form':form})
 
@@ -58,16 +85,23 @@ def registroAsignaturas(request):
 	"""Genera el form para registro de asignaturas y guarda los datos en la base de datos."""
 
 	user = request.user
+	prof = Profesor.objects.get(user = user)
+	dept = prof.departamento.codigo
 	if request.method == 'POST':
 
 		# Generamos el form
 		form = RegistrarMatForm(user, request.POST)
 
 		if form.is_valid():
+			codigo = form.cleaned_data['codigo'][0]+form.cleaned_data['codigo'][1]
 
-			form.save()
+			if (codigo == dept):
 
-			return redirect('/inicio/')
+				form.save()
+
+				return redirect('/inicio/')
+			else:
+				return render(request, 'Asignaturas/registroAsignaturas.html', {'form':form})
 		else:
 			return render(request, 'Asignaturas/registroAsignaturas.html', {'form':form})
 	else:
@@ -89,7 +123,6 @@ def registrar(request):
 				form.save()
 				username = form.cleaned_data.get('username')
 				dpto = form.cleaned_data.get('departamento')
-				# FALTA GUARDAR EL USUARIO EN profesor
 				dept = Departamento.objects.filter(codigo=dpto).first()
 				profesor = Profesor(
 					user = User.objects.get(username=username),
